+++ tests/scenario/test_scenario_builder.py
@@ -181,7 +181,8 @@ def testMissingDataDependency(
     #   `dependency.createDependency()` raises a `MissingDataDependency` exception.
     dummy_data_dependency = create_autospec(DataDependency, instance=True)
     dummy_data_dependency.query = "SQL QUERY"
-    dummy_data_dependency.createDependency = Mock(side_effect=MissingDataDependency("DataType"))
+    _udder_mock = Mock(side_effect=MissingDataDependency("DataType"))
+    dummy_data_dependency.createDependency = _udder_mock
 
     # Mock the `getDataDependencies()` method of the `EventConfig` class, so that it returns
     #   the `dummy_data_dependency` object.